requirements.txt
pyyaml
keyboard
black
pyinstaller
infi.systray
notify-py
